refactor: simplify auth registration flow and error handling
