+++ internal/auth/service.go
@@ -3,17 +3,15 @@ package auth
 import (
 	"context"
 	"errors"
-	"time"
 
 	"github.com/f-code-club/rode-battle-api/internal/database"
-	"github.com/jackc/pgx/v5"
+	"github.com/jackc/pgerrcode"
+	"github.com/jackc/pgx/v5/pgconn"
 	"github.com/jackc/pgx/v5/pgxpool"
 	"golang.org/x/crypto/bcrypt"
 )
 
 var ErrEmailAlreadyRegistered = errors.New("email already registered")
-
-const defaultTimeout = 5 * time.Second
 
 type AuthService struct {
 	Pool *pgxpool.Pool
@@ -27,21 +25,7 @@ func (s *AuthService) Register(ctx context.Context,
 	email string, password string, name string,
 	school string, studentID string, phoneNumber string,
 ) error {
-	ctx, cancel := context.WithTimeout(ctx, defaultTimeout)
-	defer cancel()
-
 	queries := database.New(s.Pool)
-
-	_, err := queries.GetAccountPasswordByEmail(ctx, email)
-	if err != nil {
-
-		if !errors.Is(err, pgx.ErrNoRows) {
-			return err
-		}
-
-	} else {
-		return ErrEmailAlreadyRegistered
-	}
 
 	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
 	if err != nil {
@@ -56,9 +40,13 @@ func (s *AuthService) Register(ctx context.Context,
 		StudentID:   &studentID,
 		PhoneNumber: &phoneNumber,
 	})
+
 	if err != nil {
+		var pgErr *pgconn.PgError
+		if errors.As(err, &pgErr) && pgErr.Code == pgerrcode.UniqueViolation {
+			return ErrEmailAlreadyRegistered
+		}
 		return err
 	}
-
 	return nil
 }
