+++ internal/database/query/account.sql
@@ -3,7 +3,7 @@ INSERT INTO accounts (email, password, name, school, student_id, phone_number)
 VALUES (@email, @password, @name, @school, @student_id, @phone_number)
 RETURNING id;
 
--- name: GetAccountPasswordByEmail :one
+-- name: GetAccountByEmail :one
 SELECT id, password
 FROM accounts
 WHERE email = @email;
